Add filesystem support
There is now a file system! The tree starts at C/ and has files within the
Desktop folder. Explorer.app will show all the files within a current directory
and can distinguish between a file and a folder. When clicking on a folder, it
will list the contents of that folder. Pretty neat!

TODO: Connect the file system with a local storage "database" that will house
actual files I'll be able to read and manipulate.